+++ CHANGES.rst
@@ -1,6 +1,11 @@
 Changes
 =======
 
+Version 0.9.1 (UNRELEASED)
+--------------------------
+
+- Changes the deletion of a workflow to automatically delete an open interactive session attached to its workspace.
+
 Version 0.9.0 (2023-01-19)
 --------------------------